refactor(ci): replace Python slack mapping with shell script integration
Replace the Python slack.py module with subprocess calls to the existing
get-slack-user-id.sh shell script. This:
- Removes duplicate user mapping maintenance
- Uses existing shell script as single source of truth
- Reduces code duplication

Changes:
- report_slack.py: Add get_slack_user_id() and get_slack_mention() functions
  that call the shell script via subprocess
- report_markdown.py: Update import to use report_slack module
- Delete slack.py module

The new implementation preserves exact behavior while delegating user mapping
to the authoritative shell script.

Partially generated by AI
Addresses code review feedback
